Добавлены скрипты с таблицами

+++ server/src/main/kotlin/com/omsk/kp/domain/model/Product.kt
@@ -7,7 +7,7 @@ import jakarta.persistence.Id
 @Entity
 data class Product(
     val name: String,
-    val price: Double
+    val price: Long
 ){
     @Id
     @GeneratedValue
